CabalTranslation.hs: patch "--force" flag into the generated unregister.sh script
If package A depends on B, and B is updated (say because of pkgrel bump) then
Pacman un-registers the current version of B before registering the new
version. At this point, ghc-pkg detects that un-registering B would break A,
and aborts the procedure. The update works anyway, because our generated shell
script script loses the error code, but we can avoid that error altogether by
giving ghc-pkg the flag '--force'. Unfortunately, there is no way to tell Cabal
that --force should be set in the generated "unregister.sh" script, so we have
to patch that flag into the command line using mad sed(1) magic.

With that change, an update of a package that's a prerequisite for another one
will look like this:

 | [...]
 |
 | (1/1) checking for file conflicts
 | unregistering tar-0.3.1.0 would break the following packages: archlinux-0.3.4 (ignoring)
 | (1/1) upgrading haskell-tar
 |
 | [...]

Basically, we tell ghc-pkg to shut up and trust that Pacman does the right
thing.