PkgBuild: drop warning about arbitrary version intervals
The archlinux library issues warnings while converting Cabal files that
specify lower and upper bounds for dependencies, like "foo > 1 && < 2",
because those restrictions cannot be expressed in a PKGBUILD file.

Now, what are those warnings good for?

That kind of version restriction is exceedingly common in Cabal files. An
attempt to convert all of Hackage results in hundreds of warnings, yet
it's not clear what anyone is supposed to do about them. There is just no
way to convert those specifications so that Pacman can process them.
Basically, we ensure that those restrictions are fulfilled long before a
PKGBUILD is ever created: i.e. by means of the PKGLIST, which specifies
the exact set of package versions that ought to be converted to ArchLinux.