Split the more generic version cabal2pkg' out of cabal2pkg.
cabal2pkg includes a hard-coded decision about the $pkgname of the generated
PKGBUILD file. The user, however, may want to override that name, so the
desired package name should to be passed as a parameter. The same applies to
the value of $pkgrel, which is currently hard-coded to "1".

The more generic function cabal2pkg' accomplishes this.