Drop inappropriate cabal2arch attribution.
The archlinux library should not make claims about who was responsible
for the generation of a PKGBUILD. It might have been cabal2arch, but it
might have been something else. This code is generic: we want it to be
re-used by other tools. Consequently, a "generated by" line should be
inserted by the tool that uses the library.