CabalTranslation.hs: minor semantic change of PKGBUILD_HASKELL_ENABLE_PROFILING
In order to enable building of profiling libraries, set the
environment variable PKGBUILD_HASKELL_ENABLE_PROFILING to any
non-empty string, i.e. "1", "0", "yes", or "no". To disable support,
undefine PKGBUILD_HASKELL_ENABLE_PROFILING (or set it to an empty
string).

These semantics feel slightly more robust than expecting the user to
specify a specific value, a number even, to turn this feature on. By
default, that environment variable is undefined, so profiling
libraries are off by default.

For those who are running chroot builds -- i.e. by means of the
makeworld script shipped in HABS --, please note that the shell
environment of the caller will not be exported into the chroot build
environment! As of now, it seems like the only way to enable
profiling in chroot builds is to add an assignment of the variable
PKGBUILD_HASKELL_ENABLE_PROFILING to /etc/makepkg.conf (in the
chroot environment).